+++ crates/subgraph/src/types/mod.rs
@@ -1,5 +1,5 @@
 pub mod order_detail;
-pub mod order_traits;
+pub mod order_detail_traits;
 pub mod orders_list;
 pub mod vault_detail;
 pub mod vaults_list;
