+++ crates/subgraph/src/types/order_detail_traits.rs
